YAY-114 refactor: transfer settings from entities to features, transfer schemas and constants. fix: replace ! with if
