+++ src/features/settings/model/constants/accountType.ts
