+++ src/features/settings/model/constants/paymentResult.ts
