+++ src/features/settings/model/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useSubscription'
