+++ src/features/settings/model/schema/index.ts
@@ -1,0 +1,1 @@
+export * from './GeneralInformationSchema'
