+++ src/features/settings/ui/DevicesTabPage/ActiveSessions/ActiveSessions.tsx
