+++ src/features/settings/ui/DevicesTabPage/CurrentDevice/CurrentDevice.tsx
