+++ src/features/settings/ui/DevicesTabPage/DevicesCard/DevicesCard.tsx
