+++ src/features/settings/ui/DevicesTabPage/DevicesTabPage.tsx
