+++ src/features/settings/ui/InfoTabPage/GeneralInormationForm/GeneralInormationForm.tsx
@@ -4,7 +4,7 @@ import { useEffect } from 'react'
 import { useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
 import z from 'zod'
-import { generalInformationSchema } from '@/entities/settings/schema/GeneralInformationSchema'
+import { generalInformationSchema } from '@/features/settings/model/schema/GeneralInformationSchema'
 import { ControlledInput, ControlledTextarea } from '@/shared/components/Controlled'
 import { ControlledDatePicker } from '@/shared/components/Controlled/ControlledDatePicker'
 import { CountriesResponse, getProfileResponse } from '@/features/settings/api/settings.types'
