+++ src/features/settings/ui/InfoTabPage/InfoTabPage.tsx
