+++ src/features/settings/ui/SubscriptionTabPage/UserAccountType/UserAccountType.tsx
@@ -1,8 +1,8 @@
 import { Typography } from '@/shared/components/Typography'
 import { RadioGroup, RadioItem } from '@/shared/components/Radio-group'
 import { Card } from '@/shared/components/Card'
-import { ACCOUNT_TYPE } from '@/entities/settings'
-import { AccountType } from '@/entities/settings/model/hooks/useSubscription'
+import { ACCOUNT_TYPE } from '@/features/settings'
+import { AccountType } from '@/features/settings/model/hooks/useSubscription'
 
 type Props = {
    value: AccountType
