+++ src/features/settings/ui/modals/AddAvatarModal/AvatarCropper.stories.tsx
